+++ Snap.Data.Mapper.Model/ExcelBinOutput/Activity/Mechanicus/MechanicusMapPointExcelConfigData.cs
@@ -7,8 +7,8 @@ public class MechanicusMapPointExcelConfigData : DataObject
     public int ID { get; set; }
 
     [JsonPropertyName("coordX")]
-    public double CoordX { get; set; }
+    public float CoordX { get; set; }
 
     [JsonPropertyName("coordY")]
-    public double CoordY { get; set; }
+    public float CoordY { get; set; }
 }
